Improvement on the policy recommendation status command
In this commit, we are adding following improvements on the policy
recommendation status command:
1. Check the ClickHouse DB for the completed jobs which are no longer on
   the API server.
2. Show the error message for the failed jobs.
3. Replace the empty state string with NEW for the jobs on NewState
